Don't swallow exceptions at all, it's just easier this way
We don't have to deal with the situation where we are returning
a non-nullable object that contains the default data and needing
to interpret that as a failed response.

It's the caller's responsibility to wrap this in exception handling
as needed; we already nicely handle and re-throw correctly the
errors encountered during the process.