Basing the decision of infinite redirects on the entire request
We were previously only using the presence of an already-seen URI as the
"Location" target in a response from the server as reason enough to
consider the redirect a cyclic/infinite redirect and abort.

This is apparently not good enough because some people at some big sites
like eBay thought it was a good reason to write code that bounces users
back and forth between the same URLs, but changing the cookie field
(amongst others) and thereby changing the actions and ultimately serving
up a response.